+++ src/components/controlled/transactions-table/components/transaction-value/transaction-value.tsx
@@ -25,14 +25,20 @@ export class TransactionValue {
         )}
 
         {this.value.showFormattedAmount && (
-          <mvx-format-amount
-            class={classNames('mr-1', { 'text-truncate': this.value.svgUrl })}
-            dataTestId={DataTestIdsEnum.transactionActionFormattedAmount}
-            isValid={true}
-            label={this.value.egldLabel}
-            valueDecimal={this.value.valueDecimal}
-            valueInteger={this.value.valueInteger}
-          />
+          <div class="amount">
+            {this.value.egldLabel && <mvx-multiversx-symbol-icon class="amount-symbol" />}
+
+            <mvx-format-amount
+              class={classNames('mr-1 mvx:text-primary', { 'text-truncate': this.value.svgUrl })}
+              dataTestId={DataTestIdsEnum.transactionActionFormattedAmount}
+              isValid={true}
+              label={this.value.egldLabel}
+              valueDecimal={this.value.valueDecimal}
+              valueInteger={this.value.valueInteger}
+              decimalClass="opacity-70"
+              labelClass="opacity-70"
+            />
+          </div>
         )}
 
         {this.value.link && (