+++ src/components/functional/toasts-list/components/transaction-toast/components/transaction-toast-content/transaction-toast-content.tsx
@@ -63,7 +63,7 @@ export class TransactionToastContent {
                   'transaction-toast-title-short': Boolean(showAmount),
                 }}
               >
-                {transaction?.action?.name || title}
+                {title}
               </h4>
               {showAmount && (
                 <mvx-format-amount